refac: move some funcs to daemon/utils
